refactor(api): 비동기 통신 코드 모듈화
